[TIME-SLOT-WEEK-VIEW-COMP][VM|DS] Refactor to remove time-slot-week-calendar component.
